refactor(store): slice up widget store and extract fedback prefs and ui stores
